feat: create use100vh hook
